Временные ряды/final project/requirements.txt
aiogram==3.3.0
yfinance 
scikit-learn==1.4.0
statsmodels==0.14.1
tensorflow==2.15.0
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4
python-dotenv==1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
